+++ lib/clearance/configuration.rb
@@ -47,6 +47,18 @@ module Clearance
     # @return [String]
     attr_accessor :mailer_sender
 
+    # Used to generate and verify password reset tokens
+    # Defaults to an `ActiveSupport::MessageVerifier` instance, initialized
+    # in concert with your application's `secret_key_base` via
+    # `Rails.application.message_verifier`.
+    # @return [#generate #verify]
+    attr_accessor :message_verifier
+
+    # Determines how long password reset emails are valid for
+    # Defaults to 15 minutes
+    # @return [Integer]
+    attr_accessor :password_reset_time_limit
+
     # The password strategy to use when authenticating and setting passwords.
     # Defaults to {Clearance::PasswordStrategies::BCrypt}.
     # @return [Module #authenticated? #password=]
@@ -93,6 +105,8 @@ module Clearance
       @cookie_name = "remember_token"
       @httponly = false
       @mailer_sender = 'reply@example.com'
+      @message_verifier = Rails.application.message_verifier("clearance")
+      @password_reset_time_limit = 15.minutes
       @redirect_url = '/'
       @routes = true
       @secure_cookie = false