+++ spec/factories.rb
@@ -7,10 +7,6 @@ FactoryGirl.define do
     email
     password 'password'
 
-    trait :with_forgotten_password do
-      confirmation_token Clearance::Token.new
-    end
-
     factory :user_with_optional_password, class: 'UserWithOptionalPassword' do
       password nil
       encrypted_password ''